requirements.txt
atproto
openai
python-dotenv
pandas